put nedb in a switchable module
